+++ lib/hive_stalker.rb
@@ -1,7 +1,7 @@
 # coding: utf-8
 
-require 'hive_stalker/hive_stalker'
 require 'hive_stalker/steam_id'
+require 'hive_stalker/client'
 
 module HiveStalker
   # Your code goes here...