+++ CHANGELOG.md
@@ -8,6 +8,8 @@ glance - what to expact from upgrading to a new version.
 
 ### Added
 
+- Basic client to access player data from Hive2 API.
+
 ### Changed
 
 ### Fixed